requirements.txt
Django==2.2.6
django-ckeditor==5.7.1
django-js-asset==1.2.2
django-mptt==0.10.0
django-password-reset==2.0
Markdown==3.1.1
mysqlclient==1.4.4
Pillow==6.2.0
Pygments==2.4.2
PyMySQL==0.9.3
pytz==2019.3
sqlparse==0.3.0
